refactor: Rename RecipeList/RecipeGrid to RecipesGrid
